Rename function and update code for `calculate_row_cols_indents`
- Renamed `buffer` function to `_buffer` because it's code that's not necessary for now.
- Made so`calculate_row_cols_indents` to directly use row, col, and indent parameters. Also modified functions to adhere to use `u8` in relevant `indents` variable.

+++ yam-dark-core/benches/count_cols.rs
@@ -565,7 +565,7 @@ fn count_batch(
     space_mask: u64,
     byte_rows: &mut [u8; 64],
     byte_cols: &mut [u8; 64],
-    indents: &mut Vec<u32>,
+    indents: &mut Vec<u8>,
 ) {
     let nl_ind0 = (newline_mask & 0xFF) as usize;
     let row0 = U8_ROW_TABLE[nl_ind0];
